move signedurlfromapi into own file. update album artwork sizing in audio player. update animation time.
