+++ app/utils/s3-signed-url.ts
@@ -29,23 +29,3 @@ export async function getSignedS3Url(key: string) {
     return new Response("Failed to fetch signed url", { status: 500 });
   }
 }
-
-// fetches signed url using a loader endpoint, since env vars are not available client-side
-export async function getSignedS3UrlFromApi(key: string): Promise<string> {
-  try {
-    const response = await fetch(
-      `/api/s3-signed-url?key=${encodeURIComponent(key)}`,
-    );
-
-    console.log(key, response, "response from signed url");
-
-    if (!response.ok) {
-      throw new Error("failed to fetch signed url");
-    }
-
-    return response.json();
-  } catch (error) {
-    console.error("Failed to fetch signed url:", error);
-    throw error;
-  }
-}
